+++ QioskAPI/Controllers/UsersController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -33,10 +34,20 @@ namespace QioskAPI.Controllers
         }
 
         // GET: api/Users
+        [Authorize]//ad
         [HttpGet]
         public async Task<ActionResult<IEnumerable<User>>> GetUsers()
         {
-            var response = await _userService.GetUsers();
+            IEnumerable<User> response;
+            var isAdmin = bool.Parse(User.Claims.FirstOrDefault(c => c.Type == "isAdmin").Value);
+            if (isAdmin)
+            {
+              response = await _userService.GetUsers();
+            }
+            else
+            {
+                return Unauthorized();
+            }
             if(response == null)
                 return BadRequest(new { message = "something went wrong in UserService" });
 
@@ -44,6 +55,7 @@ namespace QioskAPI.Controllers
         }
 
         // GET: api/Users/5
+        [Authorize]
         [HttpGet("{id}")]
         public async Task<ActionResult<User>> GetUser(int id)
         {
@@ -56,6 +68,7 @@ namespace QioskAPI.Controllers
 
         // PUT: api/Users/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [Authorize]//ad
         [HttpPut("{id}")]
         public async Task<IActionResult> PutUser(int id, User user)
         {
@@ -66,7 +79,16 @@ namespace QioskAPI.Controllers
 
             try
             {
-               await _userService.PutUser(id, user);
+                var isAdmin = bool.Parse(User.Claims.FirstOrDefault(c => c.Type == "isAdmin").Value);
+                if (isAdmin)
+                {
+
+                    await _userService.PutUser(id, user);
+                }
+                else
+                {
+                    return Unauthorized();
+                }
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -93,6 +115,7 @@ namespace QioskAPI.Controllers
         }
 
         // DELETE: api/Users/5
+        [Authorize]//ad
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(int id)
         {
@@ -101,8 +124,16 @@ namespace QioskAPI.Controllers
             {
                 return NotFound();
             }
+            var isAdmin = bool.Parse(User.Claims.FirstOrDefault(c => c.Type == "isAdmin").Value);
+            if (isAdmin)
+            {
 
-            await _userService.DeleteUser(id);
+                await _userService.DeleteUser(id);
+            }
+            else
+            {
+                return Unauthorized();
+            }
 
             return NoContent();
         }
